requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
langchain-huggingface
langchain_ollama
langchain_openai
pypdf
sentence-transformers
faiss-cpu
python-dotenv
google-generativeai  # Adicionado manualmente
cryptography       # Adicionado manualmente